Extract MediaFactory class to own file
